% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraphFromChewBBACA.r
\name{makeGraphFromChewBBACA}
\alias{makeGraphFromChewBBACA}
\title{MSTree R package}
\arguments{
\item{chewbbaca_ExtractCgMLST_output}{Full path of the file name or
a data frame.}

\item{max_allelic_difference}{numeric; This is an optional argument, can
be used to connect nodes that have a distance less than or
equal to the specified number, useful to show clusters of
isolates. The default is set to -1 which means no clustering
is applied.}

\item{verbose}{logical; set to TRUE to print on screen the progress.
default is FALSE.}
}
\value{
Graph object.
}
\description{
Make a graph object from the extracted alleles
}
\details{
Make a graph object and generate minimum spanning tree.

This function takes alleles matrix and returns a graph object
}
\examples{
my_graph <- makeGraphFromChewBBACA(system.file("extdata", "cgMLST95.csv", 
  package = "MSTree"), max_allelic_difference=-1)

}
