% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMST.r
\name{PlotMST}
\alias{PlotMST}
\title{MSTree R package}
\arguments{
\item{igraph_object}{igraph object, output of the makeGraphFromChewBBACA
function}

\item{metadata}{data frame; This is an optional argument. If given,
the function will assume the first column contains the
isolates names and will add the remaining columns as
metadata to the igraph object, useful for generating
custom coloring of the nodes i.e. Year, Source, etc.
The default is NULL}

\item{show_clustering}{logical; If the user has used a value for
max_allelic_difference when running makeGraphFromChewBBACA,
the user has the option to show the clustering of nodes based
on that threshold if set to TRUE, or FALSE to show the
minimum spanning tree without any clustering between nodes, as
if the user did not apply any clustering. The default is TRUE}

\item{MST_edges_color}{character; This argument will allow the user to set
the color of the edges that connect the nodes. The default is
"black"}

\item{cluster_edges_color}{character; This argument will allow the user to
set the color of the edges that connect the clustered nodes.
The default is "grey"}

\item{MST_edges_width}{numeric; This argument will allow the user to set
the width of the edges that connect the nodes.
The default is 0.75}

\item{cluster_edges_width}{numeric; This argument will allow the user to
set the width of the edges that connect the clustered nodes.
The default is 0.5}

\item{edge_label_size}{numeric; This argument will allow the user to set
the label size of the distances on the edges.
The default is 3}

\item{edge_label_dodge}{numeric; This argument will allow the user control
the position of the distance on the edges.
The default is 2}

\item{node_color}{character; This argument will allow the user to choose
the color of the nodes. The color can be specified by its name
or using hexadecimal representation. Also, if the metadata
argument is given, the user can specify a column name from the
metadata data frame to color the nodes i.e. "Year".
The default is "red"}

\item{node_size}{numeric; This argument will allow the user to control the
node size. The default is 3}

\item{node_label_size}{numeric; This argument will allow the user to
control the node label size. If set to zero, no labels will be
shown. The default is 0}

\item{show_legend}{logical; This argument will allow the user to
show or hide the legend of the generated plot.
The default is TRUE}

\item{title}{character; This argument will allow the user to
add a title to the generated plot.
The default is NULL}

\item{basic}{logical; This argument will allow the user to
generate a very basic minimum spanning tree layout using the
the package NetPathMiner.
The default is FALSE}
}
\value{
Graph object.
}
\description{
Using the igraph object generated from makeGraphFromChewBBACA
function, PlotMST function will plot a minimum spanning tree.
}
\details{
The function \emph{PlotMST} will make a minimum spannig tree using
the output of makeGraphFromChewBBACA function. Many arguments are
available to customize the minimum spanning tree.
}
\examples{
my_graph <- makeGraphFromChewBBACA(system.file("extdata", "cgMLST95.csv", 
   package = "MSTree"), max_allelic_difference=-1)
mst <- PlotMST(my_graph, show_clustering = TRUE, show_legend=FALSE, 
   MST_edges_color="#b97b29", node_color = "#3b17db", 
   node_label_size = 3, title = "MST")

}
